Move camera movement system from controls.rs to camera.rs
